part4 added

+++ part4/blogList/controllers/blogs.js
@@ -1,8 +1,11 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const config = require('../utils/config')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response, next) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user')
     response.json(blogs)
 })
 
@@ -17,9 +20,16 @@ blogsRouter.post('/', async (request, response, next) => {
   }else if(('title' in request.body === false) || ('url' in request.body === false)){
     response.status(400).send("Bad request")
   }else{
+    const decoded = jwt.verify(request.token, config.SECRET);
+    if(!decoded.id){
+      return response.status(401).json({error: 'invalid token'})
+    }
+    const user = await User.findById(decoded.id)
     const blog = new Blog(request.body)
     const result = await blog.save()
-    response.status(201).json(result)
+    user.blogs = user.blogs.concat(result._id)
+    await user.save()
+    response.json(result)
   }
 })
 
@@ -32,8 +42,21 @@ blogsRouter.put('/:id', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request,response,next) => {
   const id = request.params.id
+  const decoded = jwt.verify(request.token, config.SECRET);
+  if(!decoded.id){
+    return response.status(401).json({error: 'invalid token'})
+  }
+  const blog = await Blog.findById(id)
+  if(blog === null){
+    return response.status(404).json({error: 'blog does not exist'})
+  }
+  if(blog.user.toString() !== decoded.id.toString()){
+    return response.status(403).json({error: 'Unothorized operation'})
+  }
   await Blog.deleteOne({_id: id})
-  response.status(204).send("content deleted successfully")
+  response.status(201).json({
+    message: 'blog deleted successfully'
+  })
 })
 
 module.exports = blogsRouter
